# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Room v1","description":"API для взаимодействия с контроллером управления комнатами"},{"name":"User v1","description":"API для взаимодействия с контроллером управления пользователями"},{"name":"Booking v1","description":"API для взаимодействия с контроллером бронирования"},{"name":"Hotel v1","description":"API для взаимодействия с контроллером управления отелями"}],"paths":{"/api/v1/user":{"get":{"tags":["User v1","user"],"summary":"Поиск пользователя по ID","description":"Вернет UserDto объект","operationId":"findById","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}},"put":{"tags":["User v1","user"],"summary":"Изменить пользователя","description":"Вернет UserDto объект. Наличие userId обязательно","operationId":"edit","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserChangeRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}},"post":{"tags":["User v1","user"],"summary":"Создать нового пользователя","description":"Вернет UserDto объект","operationId":"createNew","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserChangeRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}},"delete":{"tags":["User v1","user"],"summary":"Удалить пользователя по Id","description":"Ничего не вернет","operationId":"deleteById","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/room":{"get":{"tags":["Room v1","room"],"summary":"Поиск комнаты по ID","description":"Вернет RoomDto объект","operationId":"findById_1","parameters":[{"name":"roomId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoomDto"}}}}}},"put":{"tags":["Room v1","room"],"summary":"Изменить комнату","description":"Вернет RoomDto объект. Наличие roomId обязательно","operationId":"edit_1","parameters":[{"name":"roomId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomChangeRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoomDto"}}}}}},"post":{"tags":["Room v1","room"],"summary":"Создать новую комнату","description":"Вернет RoomDto объект","operationId":"createNew_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomChangeRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoomDto"}}}}}},"delete":{"tags":["Room v1","room"],"summary":"Удалить комнату по Id","description":"Ничего не вернет","operationId":"deleteById_1","parameters":[{"name":"roomId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/hotel":{"get":{"tags":["hotel","Hotel v1"],"summary":"Поиск отеля по ID","description":"Вернет HotelDto объект","operationId":"findById_2","parameters":[{"name":"hotelId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HotelDto"}}}}}},"put":{"tags":["hotel","Hotel v1"],"summary":"Изменить отель","description":"Вернет HotelDto объект. Наличие hotelId обязательно","operationId":"edit_2","parameters":[{"name":"hotelId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HotelChangeRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HotelDto"}}}}}},"post":{"tags":["hotel","Hotel v1"],"summary":"Создать новый отель","description":"Вернет HotelDto объект","operationId":"createNew_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HotelChangeRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HotelDto"}}}}}},"delete":{"tags":["hotel","Hotel v1"],"summary":"Удалить отель по Id","description":"Ничего не вернет","operationId":"deleteById_2","parameters":[{"name":"hotelId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/hotel/rate":{"post":{"tags":["hotel","Hotel v1"],"summary":"Поставить оценку отелю","description":"Вернет HotelDto объект","operationId":"rate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HotelRate"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HotelDto"}}}}}}},"/api/v1/booking":{"get":{"tags":["Booking v1"],"operationId":"getAllByUserId","parameters":[{"name":"criteria","in":"query","required":true,"schema":{"$ref":"#/components/schemas/BookingSearchCriteria"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageBookingBaseDto"}}}}}},"post":{"tags":["Booking v1"],"operationId":"booking","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BookingBaseDto"}}}}}}},"/api/v1/user/findByUsername":{"get":{"tags":["User v1","user"],"summary":"Поиск пользователя по userName","description":"Вернет UserDto объект","operationId":"findByUsername","parameters":[{"name":"userName","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/api/v1/user/all":{"get":{"tags":["User v1","user"],"summary":"Получить всех пользователей","description":"Вернет Page<UserDto> объект, принимает page - номер страницы, pageSize - размер страницы","operationId":"findByCriteria","parameters":[{"name":"searchCriteria","in":"query","required":true,"schema":{"$ref":"#/components/schemas/BaseSearchCriteria"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageUserDto"}}}}}}},"/api/v1/statistic/userRegistration":{"get":{"tags":["statistic"],"summary":"Получить csv файл с регистрациями пользователей","description":"Вернет statistics.csv файл","operationId":"getUserRegistrationStatistic","responses":{"200":{"description":"OK"}}}},"/api/v1/statistic/booking":{"get":{"tags":["statistic"],"summary":"Получить csv файл с бронированием комнат","description":"Вернет statistics.csv файл","operationId":"getBookingStatistic","responses":{"200":{"description":"OK"}}}},"/api/v1/room/all":{"get":{"tags":["Room v1","room"],"summary":"Получить все комнаты","description":"Вернет Page<RoomDto> объект, принимает page - номер страницы, pageSize - размер страницы, roomId - id отеля (ОБЯЗАТЕЛЕН)","operationId":"findByCriteria_1","parameters":[{"name":"hotelSearchCriteria","in":"query","required":true,"schema":{"$ref":"#/components/schemas/RoomSearchCriteria"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageRoomDto"}}}}}}},"/api/v1/hotel/all":{"get":{"tags":["hotel","Hotel v1"],"summary":"Получить все отели","description":"Вернет Page<HotelDto> объект, принимает page - номер страницы, pageSize - размер страницы","operationId":"findByCriteria_2","parameters":[{"name":"hotelSearchCriteria","in":"query","required":true,"schema":{"$ref":"#/components/schemas/HotelSearchCriteria"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageHotelDto"}}}}}}}},"components":{"schemas":{"UserChangeRequest":{"type":"object","properties":{"username":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"role":{"uniqueItems":true,"type":"array","items":{"type":"string","enum":["ROLE_USER","ROLE_ADMIN"]}}}},"UserDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"email":{"type":"string"},"role":{"uniqueItems":true,"type":"array","items":{"type":"string","enum":["ROLE_USER","ROLE_ADMIN"]}}}},"RoomChangeRequest":{"required":["hotelId","maximumCapacity","roomDescription","roomName","roomNumber"],"type":"object","properties":{"roomName":{"type":"string"},"roomDescription":{"type":"string"},"roomNumber":{"type":"integer","format":"int32"},"maximumCapacity":{"type":"integer","format":"int32"},"hotelId":{"minimum":1,"type":"integer","format":"int64"}}},"HotelDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"hotelName":{"type":"string"},"headerAd":{"type":"string"},"cityName":{"type":"string"},"address":{"type":"string"},"distanceFromCityCenter":{"type":"integer","format":"int64"},"rating":{"type":"integer","format":"int32"},"rateCount":{"type":"integer","format":"int32"}}},"RoomDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"roomName":{"type":"string"},"roomDescription":{"type":"string"},"roomNumber":{"type":"integer","format":"int32"},"maximumCapacity":{"type":"integer","format":"int32"},"roomPrice":{"type":"number"},"hotel":{"$ref":"#/components/schemas/HotelDto"}}},"HotelChangeRequest":{"type":"object","properties":{"hotelName":{"type":"string"},"headerAd":{"type":"string"},"cityName":{"type":"string"},"address":{"type":"string"},"distanceFromCityCenter":{"type":"integer","format":"int64"}}},"HotelRate":{"required":["hotelId","rating"],"type":"object","properties":{"hotelId":{"minimum":1,"type":"integer","format":"int64"},"rating":{"maximum":5,"minimum":1,"type":"integer","format":"int32"}}},"BookingRequest":{"type":"object","properties":{"bookingFrom":{"type":"string","format":"date-time"},"bookingTo":{"type":"string","format":"date-time"},"userId":{"type":"integer","format":"int64"},"roomId":{"type":"integer","format":"int64"},"comment":{"type":"string"}}},"BookingBaseDto":{"type":"object","properties":{"bookingFrom":{"type":"string","format":"date-time"},"bookingTo":{"type":"string","format":"date-time"},"user":{"$ref":"#/components/schemas/UserDto"},"room":{"$ref":"#/components/schemas/RoomDto"},"comment":{"type":"string"}}},"BaseSearchCriteria":{"required":["page","pageSize"],"type":"object","properties":{"page":{"minimum":1,"type":"integer","description":"Номер страницы с записями. Начинается с 1","format":"int32"},"pageSize":{"minimum":1,"type":"integer","description":"Номер страницы с записями. Начинается с 1","format":"int32"},"sortField":{"type":"string","description":"Наименование поля, по которому будет производится сортировка"},"asc":{"type":"boolean","description":"Сторона сортировки ASC & DESC"}}},"PageUserDto":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"paged":{"type":"boolean"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"}}},"RoomSearchCriteria":{"required":["page","pageSize"],"type":"object","properties":{"page":{"minimum":1,"type":"integer","description":"Номер страницы с записями. Начинается с 1","format":"int32"},"pageSize":{"minimum":1,"type":"integer","description":"Номер страницы с записями. Начинается с 1","format":"int32"},"sortField":{"type":"string","description":"Наименование поля, по которому будет производится сортировка"},"asc":{"type":"boolean","description":"Сторона сортировки ASC & DESC"},"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"minPrice":{"type":"number"},"maxPrice":{"type":"number"},"maximumCapacity":{"type":"integer","format":"int32"},"bookingFrom":{"type":"string","format":"date-time"},"bookingTo":{"type":"string","format":"date-time"},"hotelId":{"type":"integer","format":"int64"}}},"PageRoomDto":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/RoomDto"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"HotelSearchCriteria":{"required":["page","pageSize"],"type":"object","properties":{"page":{"minimum":1,"type":"integer","description":"Номер страницы с записями. Начинается с 1","format":"int32"},"pageSize":{"minimum":1,"type":"integer","description":"Номер страницы с записями. Начинается с 1","format":"int32"},"sortField":{"type":"string","description":"Наименование поля, по которому будет производится сортировка"},"asc":{"type":"boolean","description":"Сторона сортировки ASC & DESC"},"id":{"type":"integer","format":"int64"},"hotelName":{"type":"string"},"hotelHeader":{"type":"string"},"city":{"type":"string"},"address":{"type":"string"},"distanceFromCityCenter":{"type":"integer","format":"int32"},"rating":{"type":"integer","format":"int32"},"ratingCount":{"type":"integer","format":"int32"}}},"PageHotelDto":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/HotelDto"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"BookingSearchCriteria":{"required":["page","pageSize"],"type":"object","properties":{"page":{"minimum":1,"type":"integer","description":"Номер страницы с записями. Начинается с 1","format":"int32"},"pageSize":{"minimum":1,"type":"integer","description":"Номер страницы с записями. Начинается с 1","format":"int32"},"sortField":{"type":"string","description":"Наименование поля, по которому будет производится сортировка"},"asc":{"type":"boolean","description":"Сторона сортировки ASC & DESC"},"bookingFrom":{"type":"string","format":"date-time"},"bookingTo":{"type":"string","format":"date-time"},"userId":{"type":"integer","format":"int64"},"roomId":{"type":"integer","format":"int64"}}},"PageBookingBaseDto":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/BookingBaseDto"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}}}}}