# src/main/java/com/finalProject/finalProjectDevOnSpring/web/dto/search/BookingSearchCriteria.java
package com.finalProject.finalProjectDevOnSpring.web.dto.search;

import lombok.Getter;
import lombok.Setter;
import org.springframework.format.annotation.DateTimeFormat;

import java.time.LocalDateTime;

@Getter
@Setter
public class BookingSearchCriteria extends BaseSearchCriteria {
    @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
    private LocalDateTime bookingFrom;
    @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
    private LocalDateTime bookingTo;
    private Long userId;
    private Long roomId;
}